Refactor unit tests in `slang_solidity` main crate
