+++ crates/solidity-v2/outputs/cargo/slang_solidity/src/tests/ast/mod.rs
@@ -1,0 +1,12 @@
+use super::fixtures;
+
+mod built_in_resolution;
+mod definition_references;
+mod identifier_path_resolution;
+mod json;
+mod linearisations;
+mod members;
+mod node_location;
+mod number_literals;
+mod typing;
+mod visitor;
