+++ crates/solidity-v2/outputs/cargo/slang_solidity/src/tests/mod.rs
@@ -1,6 +1,5 @@
 mod abi;
 mod ast;
 mod fixtures;
-mod json;
 mod thread_safety;
 mod unit;
